notifier/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "insight-notifier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "jinja2>=3.1",
    "msal>=1.28",
    "requests>=2.31",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-snapshot>=0.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
